KVPreviewer Simplified Version
I've been working on this little tool called KvPreviewer in my spare time. It's a small feature that provides real-time previews of Kivy's KV language. It's got a text input for KV code and a display area to show the outcome of the code. I've found it quite helpful for rapid prototyping and thought it might be of interest to others. Any feedback would be greatly appreciated.